✨ Implement editable section headers in TripChecklist for a more organized list! (#41)
A frontend-only solution for headers! A simple solution for adding headers in the check list, by prefixing an item with "section:" we render it as a header instead of a normal item!

The headers can be moved by drag&drop just like normal rows. Items can be easily moved between sections.

Currently you'll have to manually type in "section:" to create a new section, but we should probably find a better way to create and organize sections in the future.

Altough I don't consider this a complete solution to the issue, it is related to issue #19